+++ packages/terra-clinical-header/src/terra-dev-site/doc/example/Subheader.jsx
@@ -5,10 +5,11 @@ import Header from 'terra-clinical-header';
 const Subheader = () => (
   <div>
     <Header
-      startContent={<Button text="Consult" />}
-      title="Dr. Carl"
+      startContent={<Button aria-label="Consult, Dr. Carl" text="Consult" />}
+      text="Dr. Carl"
       endContent={<Button text="Information" />}
       isSubheader
+      level={3}
     />
   </div>
 );