+++ packages/terra-clinical-header/src/terra-dev-site/test/clinical-header/NoLevelHeader.test.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import Header from '../../../Header';
+
+export default () => <Header text="This title should not render without a provided level" id="Header" />;