+++ packages/terra-clinical-header/src/terra-dev-site/test/clinical-header/NoTitleRightContentHeader.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import classNames from 'classnames/bind';
 
 import Header from '../../../Header';
-import styles from './ClinicalHeaderCommon.test.module.scss';
+import styles from './ClinicalHeaderCommontest.module.scss';
 
 const cx = classNames.bind(styles);
 
@@ -16,5 +16,6 @@ const content = (
 export default () => (
   <Header
     endContent={content}
+    level={1}
   />
 );