+++ packages/terra-clinical-header/src/terra-dev-site/test/clinical-header/TitleHeader.test.jsx
@@ -2,4 +2,4 @@ import React from 'react';
 
 import Header from '../../../Header';
 
-export default () => <Header title="Title" id="Header" />;
+export default () => <Header text="Title" id="Header" />;