+++ packages/terra-clinical-header/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+* Changed
+  * Deprecated `title` prop. This will be replaced by the`text` prop in the next major version.
+
+* Added
+  * Added `level` prop to allow users to render different heading levels.
+  * Added `id` prop which is to be paired with `aria-labelledby` for labelling page regions.
+  * Added `text` prop to allow users to spread title HTML Attribute as customProps. This will replace the `title` prop in the next major release.
+
 ## 3.27.0 - (August 2, 2023)
 
 * Changed